[feat] #123 ChatDetail 및 Chat에 toUiModel() 작성

+++ data/src/main/java/com/mashup/data/source/remote/source/datasource/RemoteChatDataSource.kt
@@ -9,7 +9,7 @@ class RemoteChatDataSource @Inject constructor(
     private val chatService: ChatService
 ) {
 
-    suspend fun getChats(id: Int): GetChatsResponseBody {
+    suspend fun getChats(id: Long): GetChatsResponseBody {
         val response = chatService.getChats(id)
         return response.data ?: throw Exception(response.message)
     }
